Get rid of org.wildfly.swarm package in tests
