+++ testsuite/extra/src/test/java/io/smallrye/metrics/test/initialization/Initialization_Counter_ClassLevel_Test.java
@@ -15,18 +15,17 @@
  *   limitations under the License.
  */
 
-package org.wildfly.swarm.microprofile.metrics.initialization;
+package io.smallrye.metrics.test.initialization;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 
-import org.eclipse.microprofile.metrics.Counter;
 import org.eclipse.microprofile.metrics.MetricID;
 import org.eclipse.microprofile.metrics.MetricRegistry;
-import org.eclipse.microprofile.metrics.annotation.Metric;
+import org.eclipse.microprofile.metrics.Tag;
+import org.eclipse.microprofile.metrics.annotation.Counted;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -35,43 +34,51 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import io.smallrye.metrics.app.CounterImpl;
-
 @RunWith(Arquillian.class)
-public class Initialization_ProducerMethod_Test {
+public class Initialization_Counter_ClassLevel_Test {
 
     @Deployment
     public static WebArchive deployment() {
         return ShrinkWrap.create(WebArchive.class)
                 .addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml")
-                .addClasses(BeanWithMetricProducerMethod.class);
+                .addClasses(BeanWithCounter_ClassLevel.class);
     }
 
     @Inject
     MetricRegistry registry;
 
-    @Inject
-    BeanWithMetricProducerMethod bean;
-
     @Test
     public void test() {
-        MetricID metricID = new MetricID("c1");
-        assertTrue(registry.getCounters().containsKey(metricID));
-        assertEquals(111, registry.getCounters().get(metricID).getCount());
+        // metric should be created for the constructor
+        assertTrue(
+                registry.getCounters().containsKey(new MetricID("customName.BeanWithCounter_ClassLevel", new Tag("t1", "v1"))));
+
+        // metrics should be created for public methods
+        assertTrue(registry.getCounters().containsKey(new MetricID("customName.publicMethod", new Tag("t1", "v1"))));
+        assertTrue(registry.getCounters().containsKey(new MetricID("customName.publicMethod2", new Tag("t1", "v1"))));
+
+        // but not for private methods
+        assertFalse(registry.getCounters().keySet().stream()
+                .anyMatch(metricID -> metricID.getName().toLowerCase().contains("private")));
     }
 
-    public static class BeanWithMetricProducerMethod {
+    @Counted(name = "customName", tags = "t1=v1", absolute = true)
+    private static class BeanWithCounter_ClassLevel {
 
-        // a Counter that always returns 111
-        @Produces
-        @Metric(name = "c1", absolute = true)
-        public Counter producer() {
-            return new CounterImpl() {
-                @Override
-                public long getCount() {
-                    return 111;
-                }
-            };
+        public BeanWithCounter_ClassLevel() {
+
+        }
+
+        public void publicMethod() {
+
+        }
+
+        public void publicMethod2() {
+
+        }
+
+        private void privateMethod() {
+
         }
 
     }
