+++ testsuite/extra/src/test/java/io/smallrye/metrics/test/initialization/Initialization_Counter_Method_Test.java
@@ -1,4 +1,4 @@
-package org.wildfly.swarm.microprofile.metrics.initialization;
+package io.smallrye.metrics.test.initialization;
 
 import static org.junit.Assert.assertTrue;
